refactor gallery component to be reusable
