+++ pages/dao/[symbol]/gallery/index.tsx
@@ -1,21 +1,29 @@
-import { getExplorerUrl } from '@components/explorer/tools'
 import PreviousRouteBtn from '@components/PreviousRouteBtn'
-import { useMemo, useState } from 'react'
-import { PhotographIcon, PlusCircleIcon } from '@heroicons/react/outline'
-import ImgWithLoader from '@components/ImgWithLoader'
+import { useCallback, useMemo, useState } from 'react'
+import { PlusCircleIcon } from '@heroicons/react/outline'
 import Modal from '@components/Modal'
 import { LinkButton } from '@components/Button'
-import { MdScheduleSend } from 'react-icons/md'
-import useLegacyConnectionContext from '@hooks/useLegacyConnectionContext'
 import { useRealmDigitalAssetsQuery } from '@hooks/queries/digitalAssets'
 import SendNft from '@components/SendNft'
 import { PublicKey } from '@solana/web3.js'
 import { SUPPORT_CNFTS } from '@constants/flags'
 import useFindGovernanceByTreasury from '@hooks/useFindGovernanceByTreasury'
 import DepositNFTFromWallet from '@components/TreasuryAccount/DepositNFTFromWallet'
+import NFTGallery from '@components/NFTGallery'
+
+const AllNFTsGallery = (
+  props: Omit<Parameters<typeof NFTGallery>[0], 'nfts'>
+) => {
+  const { data: nftsDAS } = useRealmDigitalAssetsQuery()
+  const DASnftsFlat = useMemo(
+    () =>
+      nftsDAS?.flat().filter((x) => SUPPORT_CNFTS || !x.compression.compressed),
+    [nftsDAS]
+  )
+  return <NFTGallery nfts={DASnftsFlat} {...props} />
+}
 
 const Gallery = () => {
-  const connection = useLegacyConnectionContext()
   const [openNftDepositModal, setOpenNftDepositModal] = useState(false)
   const [openSendNftsModal, setOpenSendNftsModal] = useState(false)
   const [
@@ -29,14 +37,18 @@ const Gallery = () => {
     setOpenSendNftsModal(false)
   }
 
-  const { data: nftsDAS } = useRealmDigitalAssetsQuery()
-  const DASnftsFlat = useMemo(
-    () =>
-      nftsDAS?.flat().filter((x) => SUPPORT_CNFTS || !x.compression.compressed),
-    [nftsDAS]
-  )
-
   const findGovernanceByTreasury = useFindGovernanceByTreasury()
+
+  // x is an nft object from the DAS api
+  const onClickSendNft = useCallback(
+    async (x: any) => {
+      const owner = new PublicKey(x.ownership.owner)
+      const governance = (await findGovernanceByTreasury(owner)) ?? owner
+      setSelectedNftAndItsGovernance([new PublicKey(x.id), governance])
+      setOpenSendNftsModal(true)
+    },
+    [findGovernanceByTreasury]
+  )
 
   return (
     <div className="bg-bkg-2 rounded-lg p-4 md:p-6">
@@ -109,63 +121,7 @@ const Gallery = () => {
               ))}
             </Select> */}
           </div>
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 grid-flow-row gap-6">
-            {DASnftsFlat === undefined ? (
-              <>
-                <div className="animate-pulse bg-bkg-3 col-span-1 h-48 rounded-lg" />
-                <div className="animate-pulse bg-bkg-3 col-span-1 h-48 rounded-lg" />
-                <div className="animate-pulse bg-bkg-3 col-span-1 h-48 rounded-lg" />
-                <div className="animate-pulse bg-bkg-3 col-span-1 h-48 rounded-lg" />
-              </>
-            ) : DASnftsFlat?.length ? (
-              DASnftsFlat.map((x) => (
-                <div
-                  key={x.id}
-                  className="relative group bg-bkg-4 col-span-1 flex items-center justify-center rounded-lg filter drop-shadow-xl"
-                >
-                  <a
-                    className="bg-bkg-2 cursor-pointer default-transition h-full w-full rounded-md border border-transparent transform scale-90 group-hover:scale-95 group-hover:opacity-50"
-                    href={
-                      connection.endpoint && x.id
-                        ? getExplorerUrl(connection.cluster, x.id)
-                        : ''
-                    }
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    onClick={(e) => e.stopPropagation()}
-                  >
-                    <ImgWithLoader
-                      className="h-full w-full"
-                      src={
-                        x.content.files[0]?.cdn_uri ?? x.content.files[0]?.uri
-                      }
-                    />
-                  </a>
-                  <button
-                    className="hidden group-hover:block absolute w-20 h-20 items-center justify-center flex-auto text-primary-light"
-                    onClick={async () => {
-                      const owner = new PublicKey(x.ownership.owner)
-                      const governance =
-                        (await findGovernanceByTreasury(owner)) ?? owner
-                      setSelectedNftAndItsGovernance([
-                        new PublicKey(x.id),
-                        governance,
-                      ])
-                      setOpenSendNftsModal(true)
-                    }}
-                  >
-                    <div className="bg-white rounded-full flex items-center justify-center h-full w-full p-2 hover:opacity-75">
-                      <MdScheduleSend className="h-full w-full p-3" />
-                    </div>
-                  </button>
-                </div>
-              ))
-            ) : (
-              <div className="col-span-4 text-fgd-3 flex flex-col items-center">
-                <PhotographIcon className="opacity-5 w-56 h-56" />
-              </div>
-            )}
-          </div>
+          <AllNFTsGallery onClickSendNft={onClickSendNft} />
         </div>
       </div>
       {openNftDepositModal && (
